+++ tests/TestClasses/NestedChildForm.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ansien\RapidFormBundle\Tests\TestClasses;
+
+use Ansien\RapidFormBundle\Attribute\Form;
+use Ansien\RapidFormBundle\Attribute\FormField;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+
+#[Form]
+class NestedChildForm
+{
+    #[FormField(TextType::class)]
+    public string $test = '123';
+}