+++ tests/TestClasses/TestForm.php
@@ -11,7 +11,6 @@ use DateTimeImmutable;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\DateType;
 use Symfony\Component\Form\Extension\Core\Type\NumberType;
-use Symfony\Component\Validator\Constraints as Assert;
 
 #[Form]
 class TestForm
@@ -26,22 +25,18 @@ class TestForm
         'choices' => [CallbackType::VALUE, 'enabledCurrencyChoices'],
         'choice_label' => [CallbackType::FUNCTION, 'getCurrencyLabelCb'],
     ])]
-    #[Assert\NotBlank]
     public ?string $currency = null;
 
     #[FormField(NumberType::class, [
         'scale' => 4,
         'input' => 'string',
     ])]
-    #[Assert\NotBlank]
-    #[Assert\Range(min: -1000000, max: 1000000)]
     public ?string $rate = null;
 
     #[FormField(DateType::class, [
         'widget' => 'single_text',
         'input' => 'datetime_immutable',
     ])]
-    #[Assert\NotBlank]
     public ?DateTimeImmutable $date = null;
 
     // endregion