+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `rapid-form-bundle` will be documented in this file
 
+## 1.1.0 - 2020-03-13
+- Added support for collection types
+- Added support for embeddable (nested) forms.
+
 ## 1.0.6 - 2020-03-09
 - Fixed bug in options callback transformer